Fix: TS Issues

+++ src/models/user.model.ts
@@ -1,20 +1,22 @@
-import * as mongoose from 'mongoose';
+import { model, Schema, Document, Types } from 'mongoose';
 import validator from 'validator';
 import log from '../utils/log';
 
 import { serverIo, games } from '../index';
 
-import Room from './room.model';
+import { Room, IRoom, IRoomDoc } from './room.model';
 
-export interface IUser extends mongoose.Document {
+interface IUser {
   username: string;
   iconId: number;
-  colorID: number;
+  colorId: number;
   socketId: string;
   isAdmin: boolean;
 }
 
-const userSchema: mongoose.Schema = new mongoose.Schema({
+interface IUserDoc extends IUser, Document {}
+
+const UserSchemaFields: Record<keyof IUser, any> = {
   username: {
     type: String,
     required: [true, 'Enter  a username'],
@@ -42,9 +44,11 @@ const userSchema: mongoose.Schema = new mongoose.Schema({
     type: Boolean,
     required: true,
   },
-});
+};
 
-userSchema.virtual('room', {
+const UserSchema = new Schema(UserSchemaFields);
+
+UserSchema.virtual('room', {
   ref: 'User',
   // What field is referenced when used by the other 'object'
   // This is like the foreign key
@@ -54,47 +58,49 @@ userSchema.virtual('room', {
   foreignField: 'admin',
 });
 
-userSchema.pre('deleteOne', { document: true, query: false }, async function () {
+UserSchema.pre('deleteOne', { document: true, query: false }, async function () {
   const user = this;
 
-  const room = await (
-    await Room.findOne({ users: user }).populate('users')
-  ).execPopulate();
+  const room: IRoomDoc | null = await Room.findOne({ users: user })
+    .populate('users')
+    .exec();
 
-  if (room.users.length <= 1) {
-    room
-      .deleteOne()
-      .then(() => {
-        log.success('Last user in room left, deleted room');
-      })
-      .catch((error) => {
-        throw new Error(error);
+  if (room) {
+    if (room.users.length <= 1) {
+      room
+        .deleteOne()
+        .then(() => {
+          log.success('Last user in room left, deleted room');
+        })
+        .catch((error: any) => {
+          throw new Error(error);
+        });
+    } else {
+      // Filtering out the player that is leaving
+      const newUserList: any = (room.users as Array<IUser>).filter((currentUser: any) => {
+        return currentUser._id.toString() !== user._id.toString();
       });
-  } else {
-    // Filtering out the player that is leaving
-    const newUserList: any = room.users.filter((currentUser: any) => {
-      return currentUser._id.toString() !== user._id.toString();
-    });
-    room.users = newUserList;
-    // Setting the new admin
-    if (user._id.toString() === room.admin.toString()) {
-      newUserList[0].isAdmin = true;
-      room.admin = newUserList[0];
-      newUserList[0].save();
-    }
-    await room.save();
-    serverIo.to(room.name).emit('updatePlayers', newUserList);
+      room.users = newUserList;
+      // Setting the new admin
+      if (user._id.toString() === room.admin.toString()) {
+        newUserList[0].isAdmin = true;
+        room.admin = newUserList[0];
+        newUserList[0].save();
+      }
+      await room.save();
+      serverIo.to(room.name).emit('updatePlayers', newUserList);
 
-    // Setting the player that left as finished so other players don't
-    //    have to wait for that player's timers to finish
-    //  We want the player to still show up as participation in the score
-    //    which is why they aren't simply deleted
-    if (games[room._id]) {
-      games[room.id].handlePlayerLeaving(user._id);
+      // Setting the player that left as finished so other players don't
+      //    have to wait for that player's timers to finish
+      //  We want the player to still show up as participation in the score
+      //    which is why they aren't simply deleted
+      if (games[room._id]) {
+        games[room.id].handlePlayerLeaving(user._id);
+      }
     }
   }
 });
 
-const User = mongoose.model<IUser>('User', userSchema);
+const User = model<IUserDoc>('User', UserSchema);
 
-export default User;
+export { User, IUser };
